YouTube import: Handle excessive traffic error from YouTube
When issuing too many queries to YouTube, it will report a quota error
and we need to stop importing. All videos imported up to that point are
stored in the database and the user sees a message explaining the
situation and can continue the import a few minutes later.

References:
http://code.google.com/apis/youtube/2.0/developers_guide_protocol_error_responses.html#Quota_errors